Add DesignExplorer handler
